module_ocr/requirements.txt

fastapi==0.101.0
    # via -r requirements.in
paddleocr==2.7.0.0
    # via -r requirements.in
paddlepaddle==2.5.1
    # via -r requirements.in
python-multipart==0.0.6
    # via -r requirements.in
uvicorn==0.23.2
    # via -r requirements.in
requests==2.31.0
    # via -r requirements.in
numpy==1.23.5
    # via -r requirements.in
pytest

language_tool_python

python-bidi

arabic_reshaper
 
opencv-python

imgaug